Turning 90 degress (could work)
Worked on using 90 degree turns for exact align. We are using pigeon and we have the basics down. Right now, the issues include having an absolute turn which would only work for certain headings, calculating dtheta after sending it to the drivetrain so it won't turn the first time, if the current yaw is around +pi or -pi, the isFinished() command my trigger too early, and the small motor commands might not move the robot if it has a deadzone.

:)